requirements.txt
kafka-python-ng==2.2.2
pyspark==3.5.1
